separate requalification tests
